organized the files and added new dependencies
