+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,8 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     package="com.example.fatapp">
 
+
+
     <!-- io.flutter.app.FlutterApplication is an android.app.Application that
          calls FlutterMain.startInitialization(this); in its onCreate method.
          In most cases you can leave this as-is, but you if you want to provide
@@ -10,6 +12,8 @@
         android:name="io.flutter.app.FlutterApplication"
         android:label="fatapp"
         android:icon="@mipmap/ic_launcher">
+        <activity android:name="com.apptreesoftware.barcodescan.BarcodeScannerActivity"/>
+        <uses-permission android:name="android.permission.CAMERA" />
         <activity
             android:name=".MainActivity"
             android:launchMode="singleTop"
