+++ android/build.gradle
@@ -1,4 +1,6 @@
 buildscript {
+    ext.kotlin_version = '1.2.31'
+
     repositories {
         google()
         jcenter()
@@ -6,6 +8,7 @@ buildscript {
 
     dependencies {
         classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
