--- lib/pages/login2.dart
@@ -1,76 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter_screenutil/flutter_screenutil.dart';
-
-class LoginForm extends StatelessWidget {
-  static String tag = 'login-page';
-
-  @override
-  Widget build(BuildContext context) {
-    return new Container(
-      width: double.infinity,
-      height: 500.0,
-      decoration: BoxDecoration(
-          color: Colors.white,
-          borderRadius: BorderRadius.circular(8.0),
-          boxShadow: [
-            BoxShadow(
-                color: Colors.black12,
-                offset: Offset(0.0, 15.0),
-                blurRadius: 15.0),
-            BoxShadow(
-                color: Colors.black12,
-                offset: Offset(0.0, -10.0),
-                blurRadius: 10.0),
-          ]),
-    child: Padding(
-        padding: EdgeInsets.only(left: 16.0, right: 16.0, top: 16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: <Widget>[
-            Text("Login",
-                style: TextStyle(
-                    fontSize: 45.0,
-                    letterSpacing: .6)),
-            SizedBox(
-              height: 30.0,
-            ),
-            Text("Username",
-                style: TextStyle(
-                    fontSize: 26.0)),
-            TextField(
-              decoration: InputDecoration(
-                  hintText: "username",
-                  hintStyle: TextStyle(color: Colors.grey, fontSize: 12.0)),
-            ),
-            SizedBox(
-              height: 30.0,
-            ),
-            Text("PassWord",
-                style: TextStyle(
-                    fontSize: 26.0)),
-            TextField(
-              obscureText: true,
-              decoration: InputDecoration(
-                  hintText: "Password",
-                  hintStyle: TextStyle(color: Colors.grey, fontSize: 12.0)),
-            ),
-            SizedBox(
-              height: 35.0,
-            ),
-            Row(
-              mainAxisAlignment: MainAxisAlignment.end,
-              children: <Widget>[
-                Text(
-                  "Forgot Password?",
-                  style: TextStyle(
-                      color: Colors.blue,
-                      fontSize: 28.0),
-                )
-              ],
-            )
-          ],
-        ),
-      ),
-    );
-  }
-}
