+++ lib/pages/views/qrCodeScan.dart
