+++ lib/pages/views/register.dart
